Update annotation_guidlines.md
I've written down what I think we're actually doing here with all the revisions.
I propose we show this to Adam on the next meeting and as him if this is the kind of description of the Golden Standard he meant. 
Furthermore, I would also add certain other things about the Golden Standard, namely details like - the kinds of linguistic structure we do and don't take into account, how many reviews we've treated as the base case for the GS (50). Also, we should refer to the user contribution to the GS (the pop-up questions answering which will save us work) - this also needs covering.